Reworking scan_BLE
