+++ extras/reset_mbients.py
@@ -7,7 +7,8 @@ Created on Fri May 27 14:49:22 2022
 
 from mbientlab.metawear import MetaWear, libmetawear
 from time import sleep
-from neurobooth_os.iout.lsl_streamer import scann_BLE
+from neurobooth_os.iout.mbient import scan_BLE
+from neurobooth_os.logging import make_default_logger
 
 
 def connect(device):
@@ -56,17 +57,23 @@ def reset_dev(MAC):
     return True
 
 
-if __name__ == "__main__":
+def main():
+    logger = make_default_logger()
+
     macs = {
         "Mbient_LH_2": "E8:95:D6:F7:39:D2",
         "Mbient_RH_2": "FE:07:3E:37:F5:9C",
         "Mbient_RF_2": "E5:F6:FB:6D:11:8A",
         "Mbient_LF_2": "DA:B0:96:E4:7F:A3",
         "Mbient_BK_1": "D7:B0:7E:C2:A1:23",
-        }
-    
+    }
+
     print('resetting mbients (will take ~ 1 min)...')
-    scann_BLE(2)
+    devices = scan_BLE(2)
+    logger.info(f'Identified {len(devices)} devices')
+    for k, v in devices.items():
+        logger.debug(f'Device {k} Address: {v}')
+
     for k, v in macs.items():
         success = reset_dev(v)
         if not success:
@@ -75,3 +82,7 @@ if __name__ == "__main__":
             print(f"Success in resetting {k} {v}")
 
     sleep(60)
+
+
+if __name__ == "__main__":
+    main()
